feat: Encryption algorithm separation
